+++ src/UserRepository/HtpasswdFactory.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-authentication for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (https://www.zend.com)
+ * @copyright Copyright (c) 2017-2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-authentication/blob/master/LICENSE.md New BSD License
  */
 
@@ -11,6 +11,7 @@ namespace Zend\Expressive\Authentication\UserRepository;
 
 use Psr\Container\ContainerInterface;
 use Zend\Expressive\Authentication\Exception;
+use Zend\Expressive\Authentication\UserInterface;
 
 class HtpasswdFactory
 {
@@ -26,7 +27,9 @@ class HtpasswdFactory
                 Htpasswd::class
             ));
         }
-
-        return new Htpasswd($htpasswd);
+        return new Htpasswd(
+            $htpasswd,
+            $container->get(UserInterface::class)
+        );
     }
 }