+++ src/UserRepository/PdoDatabase.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-authentication for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (https://www.zend.com)
+ * @copyright Copyright (c) 2017-2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-authentication/blob/master/LICENSE.md New BSD License
  */
 
@@ -17,12 +17,11 @@ use Zend\Expressive\Authentication\UserRepositoryInterface;
 
 /**
  * Adapter for PDO database
- * It supports only bcrypt hash password for security reason
+ *
+ * It supports only bcrypt password hashing for security reasons.
  */
 class PdoDatabase implements UserRepositoryInterface
 {
-    use UserTrait;
-
     /**
      * @var PDO
      */
@@ -33,10 +32,27 @@ class PdoDatabase implements UserRepositoryInterface
      */
     private $config;
 
-    public function __construct(PDO $pdo, array $config)
-    {
+    /**
+     * @var callable
+     */
+    private $userFactory;
+
+    public function __construct(
+        PDO $pdo,
+        array $config,
+        callable $userFactory
+    ) {
         $this->pdo = $pdo;
         $this->config = $config;
+
+        // Provide type safety for the composed user factory.
+        $this->userFactory = function (
+            string $identity,
+            array $roles = [],
+            array $details = []
+        ) use ($userFactory) : UserInterface {
+            return $userFactory($identity, $roles, $details);
+        };
     }
 
     /**
@@ -66,15 +82,23 @@ class PdoDatabase implements UserRepositoryInterface
             return null;
         }
 
-        return password_verify($password, $result->{$this->config['field']['password']})
-            ? $this->generateUser($credential, $this->getRolesFromUser($credential))
-            : null;
+        if (password_verify($password, $result->{$this->config['field']['password']})) {
+            return ($this->userFactory)(
+                $credential,
+                $this->getUserRoles($credential),
+                $this->getUserDetails($credential)
+            );
+        }
+        return null;
     }
 
     /**
-     * {@inheritDoc}
+     * Get the user roles if present.
+     *
+     * @param string $identity
+     * @return string[]
      */
-    public function getRolesFromUser(string $identity) : array
+    protected function getUserRoles(string $identity) : array
     {
         if (! isset($this->config['sql_get_roles'])) {
             return [];
@@ -82,7 +106,7 @@ class PdoDatabase implements UserRepositoryInterface
 
         if (false === strpos($this->config['sql_get_roles'], ':identity')) {
             throw new Exception\InvalidConfigException(
-                'The sql_get_roles configuration setting must include a :identity parameter'
+                'The sql_get_roles configuration setting must include an :identity parameter'
             );
         }
 
@@ -110,5 +134,44 @@ class PdoDatabase implements UserRepositoryInterface
             $roles[] = $role[0];
         }
         return $roles;
+    }
+
+    /**
+     * Get the user details if present.
+     *
+     * @param string $identity
+     * @return string[]
+     */
+    protected function getUserDetails(string $identity) : array
+    {
+        if (! isset($this->config['sql_get_details'])) {
+            return [];
+        }
+
+        if (false === strpos($this->config['sql_get_details'], ':identity')) {
+            throw new Exception\InvalidConfigException(
+                'The sql_get_details configuration setting must include a :identity parameter'
+            );
+        }
+
+        try {
+            $stmt = $this->pdo->prepare($this->config['sql_get_details']);
+        } catch (PDOException $e) {
+            throw new Exception\RuntimeException(sprintf(
+                'Error preparing retrieval of user details: %s',
+                $e->getMessage()
+            ));
+        }
+        if (false === $stmt) {
+            throw new Exception\RuntimeException(sprintf(
+                'Error preparing retrieval of user details: unknown error'
+            ));
+        }
+        $stmt->bindParam(':identity', $identity);
+
+        if (! $stmt->execute()) {
+            return [];
+        }
+        return $stmt->fetch(PDO::FETCH_ASSOC);
     }
 }