+++ test/UserRepository/PdoDatabaseTest.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-authentication for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (https://www.zend.com)
+ * @copyright Copyright (c) 2017-2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-authentication/blob/master/LICENSE.md New BSD License
  */
 
@@ -11,6 +11,8 @@ namespace ZendTest\Expressive\Authentication\UserRepository;
 
 use PDO;
 use PHPUnit\Framework\TestCase;
+use Prophecy\Argument;
+use Zend\Expressive\Authentication\DefaultUser;
 use Zend\Expressive\Authentication\Exception\InvalidConfigException;
 use Zend\Expressive\Authentication\Exception\RuntimeException;
 use Zend\Expressive\Authentication\UserInterface;
@@ -19,9 +21,20 @@ use Zend\Expressive\Authentication\UserRepository\PdoDatabase;
 
 class PdoDatabaseTest extends TestCase
 {
+    protected function setUp()
+    {
+        $this->userFactory = function ($identity, $roles, $details) {
+            return new DefaultUser($identity, $roles, $details);
+        };
+    }
+
     public function testConstructor()
     {
-        $pdoDatabase = new PdoDatabase(new PDO('sqlite::memory:'), []);
+        $pdoDatabase = new PdoDatabase(
+            new PDO('sqlite::memory:'),
+            [],
+            $this->userFactory
+        );
         $this->assertInstanceOf(UserRepositoryInterface::class, $pdoDatabase);
     }
 
@@ -39,7 +52,11 @@ class PdoDatabaseTest extends TestCase
     public function testAuthenticate()
     {
         $pdo = new PDO('sqlite:'. __DIR__ . '/../TestAssets/pdo.sqlite');
-        $pdoDatabase = new PdoDatabase($pdo, $this->getConfig());
+        $pdoDatabase = new PdoDatabase(
+            $pdo,
+            $this->getConfig(),
+            $this->userFactory
+        );
 
         $user = $pdoDatabase->authenticate('test', 'password');
         $this->assertInstanceOf(UserInterface::class, $user);
@@ -52,7 +69,11 @@ class PdoDatabaseTest extends TestCase
         $config = $this->getConfig();
         $config['field']['identity'] = 'foo'; // mistake in the configuration
 
-        $pdoDatabase = new PdoDatabase($pdo, $config);
+        $pdoDatabase = new PdoDatabase(
+            $pdo,
+            $config,
+            $this->userFactory
+        );
         $this->expectException(RuntimeException::class);
         $user = $pdoDatabase->authenticate('test', 'password');
     }
@@ -60,7 +81,11 @@ class PdoDatabaseTest extends TestCase
     public function testAuthenticateInvalidUserPassword()
     {
         $pdo = new PDO('sqlite:'. __DIR__ . '/../TestAssets/pdo.sqlite');
-        $pdoDatabase = new PdoDatabase($pdo, $this->getConfig());
+        $pdoDatabase = new PdoDatabase(
+            $pdo,
+            $this->getConfig(),
+            $this->userFactory
+        );
 
         $user = $pdoDatabase->authenticate('test', 'foo');
         $this->assertNull($user);
@@ -69,7 +94,11 @@ class PdoDatabaseTest extends TestCase
     public function testAuthenticateInvalidUsername()
     {
         $pdo = new PDO('sqlite:'. __DIR__ . '/../TestAssets/pdo.sqlite');
-        $pdoDatabase = new PdoDatabase($pdo, $this->getConfig());
+        $pdoDatabase = new PdoDatabase(
+            $pdo,
+            $this->getConfig(),
+            $this->userFactory
+        );
 
         $user = $pdoDatabase->authenticate('invalidusername', 'password');
         $this->assertNull($user);
@@ -80,12 +109,16 @@ class PdoDatabaseTest extends TestCase
         $pdo = new PDO('sqlite:'. __DIR__ . '/../TestAssets/pdo_role.sqlite');
         $config = $this->getConfig();
         $config['sql_get_roles'] = 'SELECT role FROM user WHERE username = :identity';
-        $pdoDatabase = new PdoDatabase($pdo, $config);
+
+        $pdoDatabase = new PdoDatabase(
+            $pdo,
+            $config,
+            $this->userFactory
+        );
 
         $user = $pdoDatabase->authenticate('test', 'password');
         $this->assertInstanceOf(UserInterface::class, $user);
-        $this->assertEquals('test', $user->getIdentity());
-        $this->assertEquals(['admin'], $user->getUserRoles());
+        $this->assertEquals(['admin'], $user->getRoles());
     }
 
     public function testAuthenticateWithRoles()
@@ -93,12 +126,54 @@ class PdoDatabaseTest extends TestCase
         $pdo = new PDO('sqlite:'. __DIR__ . '/../TestAssets/pdo_roles.sqlite');
         $config = $this->getConfig();
         $config['sql_get_roles'] = 'SELECT role FROM user_role WHERE username = :identity';
-        $pdoDatabase = new PdoDatabase($pdo, $config);
+
+
+        $pdoDatabase = new PdoDatabase(
+            $pdo,
+            $config,
+            $this->userFactory
+        );
+        $user = $pdoDatabase->authenticate('test', 'password');
+        $this->assertInstanceOf(UserInterface::class, $user);
+        $this->assertEquals(['user', 'admin'], $user->getRoles());
+    }
+
+    public function testAuthenticateWithDetails()
+    {
+        $pdo = new PDO('sqlite:'. __DIR__ . '/../TestAssets/pdo_role.sqlite');
+        $config = $this->getConfig();
+        $config['sql_get_details'] = 'SELECT email FROM user WHERE username = :identity';
+
+        $pdoDatabase = new PdoDatabase(
+            $pdo,
+            $config,
+            $this->userFactory
+        );
 
         $user = $pdoDatabase->authenticate('test', 'password');
         $this->assertInstanceOf(UserInterface::class, $user);
-        $this->assertEquals('test', $user->getIdentity());
-        $this->assertEquals(['user', 'admin'], $user->getUserRoles());
+        $this->assertEquals(['email' => 'test@foo.com'], $user->getDetails());
+        $this->assertEquals('test@foo.com', $user->getDetail('email'));
+    }
+
+    public function testAuthenticateWithRolesAndDetails()
+    {
+        $pdo = new PDO('sqlite:'. __DIR__ . '/../TestAssets/pdo_roles.sqlite');
+        $config = $this->getConfig();
+        $config['sql_get_roles'] = 'SELECT role FROM user_role WHERE username = :identity';
+        $config['sql_get_details'] = 'SELECT email FROM user WHERE username = :identity';
+
+        $pdoDatabase = new PdoDatabase(
+            $pdo,
+            $config,
+            $this->userFactory
+        );
+
+        $user = $pdoDatabase->authenticate('test', 'password');
+        $this->assertInstanceOf(UserInterface::class, $user);
+        $this->assertEquals(['email' => 'test@foo.com'], $user->getDetails());
+        $this->assertEquals('test@foo.com', $user->getDetail('email'));
+        $this->assertEquals(['user', 'admin'], $user->getRoles());
     }
 
     public function testAuthenticateWithRoleRuntimeError()
@@ -107,31 +182,44 @@ class PdoDatabaseTest extends TestCase
         $config = $this->getConfig();
         // add a mistake in the configuration
         $config['sql_get_roles'] = 'SELECT role FROM user WHERE foo = :identity';
-        $pdoDatabase = new PdoDatabase($pdo, $config);
+        $pdoDatabase = new PdoDatabase(
+            $pdo,
+            $config,
+            $this->userFactory
+        );
 
         $this->expectException(RuntimeException::class);
         $user = $pdoDatabase->authenticate('test', 'password');
     }
 
-    public function testGetRolesFromUserWithEmptySql()
+    public function testAuthenticateWithEmptySql()
     {
         $pdo = new PDO('sqlite:'. __DIR__ . '/../TestAssets/pdo_roles.sqlite');
         $config = $this->getConfig();
 
-        $pdoDatabase = new PdoDatabase($pdo, $config);
-        $roles = $pdoDatabase->getRolesFromUser('foo');
-        $this->assertEmpty($roles);
+        $pdoDatabase = new PdoDatabase(
+            $pdo,
+            $config,
+            $this->userFactory
+        );
+        $user = $pdoDatabase->authenticate('test', 'password');
+        $this->assertInstanceOf(UserInterface::class, $user);
+        $this->assertEquals('test', $user->getIdentity());
     }
 
-    public function testGetRolesFromUserWithNoIdentityParam()
+    public function testAuthenticateWithNoIdentityParam()
     {
         $pdo = new PDO('sqlite:'. __DIR__ . '/../TestAssets/pdo_roles.sqlite');
         $config = $this->getConfig();
         $config['sql_get_roles'] = 'SELECT role FROM user_role';
 
-        $pdoDatabase = new PdoDatabase($pdo, $config);
+        $pdoDatabase = new PdoDatabase(
+            $pdo,
+            $config,
+            $this->userFactory
+        );
 
         $this->expectException(InvalidConfigException::class);
-        $roles = $pdoDatabase->getRolesFromUser('foo');
+        $user = $pdoDatabase->authenticate('test', 'password');
     }
 }